Seperate rebuild service from build service
